Revert "Issue 137"
