+++ src/apresentacao/ListaProfessores.form
@@ -6,7 +6,7 @@
     <Property name="title" type="java.lang.String" value="Lista de Professores"/>
     <Property name="alwaysOnTop" type="boolean" value="true"/>
     <Property name="cursor" type="java.awt.Cursor" editor="org.netbeans.modules.form.editors2.CursorEditor">
-      <Color id="Cursor Padr&#xe3;o"/>
+      <Color id="Default Cursor"/>
     </Property>
     <Property name="name" type="java.lang.String" value="frm_ListaProfessores" noResource="true"/>
     <Property name="resizable" type="boolean" value="false"/>
