+++ src/servico/ProfessorService.java
@@ -7,7 +7,6 @@ package servico;
 
 import datamapper.AusenciaJpaController;
 import datamapper.ProfessorJpaController;
-import datamapper.exceptions.NonexistentEntityException;
 import dominio.Aula;
 import dominio.Ausencia;
 import dominio.Professor;
@@ -53,7 +52,7 @@ public class ProfessorService {
         return modelos;
     }
     
-    public List<ProfessorModel> listarProfessoresCompativeisComAusenteNoPeriodo(String codigoAusencia) throws NonexistentEntityException{
+    public List<ProfessorModel> listarProfessoresCompativeisComAusenteNoPeriodo(String codigoAusencia){
         
         Ausencia ausência = ausenciasRepository.findAusencia(new Long(codigoAusencia ));
