+++ test/servico/ProfessorServiceTest.java
@@ -84,7 +84,7 @@ public class ProfessorServiceTest {
     }
     
     @Test
-    public void testeDeveListarProfessoresPossiveisDadoUmPeriodoDeAusenciaEAulasDeProfessorAusente() throws NonexistentEntityException{
+    public void testeDeveListarProfessoresPossiveisDadoUmPeriodoDeAusenciaEAulasDeProfessorAusente(){
         
         //Ausência numa segunda feira.
         DateTime inicio = new DateTime(2013, 11, 18, 0, 0);
