+++ test/servico/cst04Test.java
@@ -57,7 +57,7 @@ public class cst04Test {
     }
     
     @Test
-    public void testeEfetuarAlocação_Confirmar() throws ParseException, NonexistentEntityException
+    public void testeEfetuarAlocação_Confirmar() throws ParseException
     {
         LoginService loginService = new LoginService();
         assertTrue(loginService.verificarUsuarioESenha("Administrador", "123456"));
