+++ test/servico/cst07Test.java
@@ -91,7 +91,7 @@ public class cst07Test {
     }
     
     @Test
-    public void testeEfetuarRealocação_Confirmar() throws ParseException, NonexistentEntityException
+    public void testeEfetuarRealocação_Confirmar() throws ParseException
     {
         LoginService loginService = new LoginService();
         assertTrue(loginService.verificarUsuarioESenha("Administrador", "123456"));
